Support noMavenJavadoc

+++ mx.py
@@ -11415,34 +11415,40 @@ def _maven_deploy_dists(dists, versionGetter, repo, settingsXml,
                 if dist.isJARDistribution():
                     javadocPath = None
                     if generateJavadoc:
-                        projects = [p for p in dist.archived_deps() if p.isJavaProject()]
-                        tmpDir = tempfile.mkdtemp(prefix='mx-javadoc')
-                        javadocArgs = ['--base', tmpDir, '--unified', '--projects', ','.join((p.name for p in projects))]
-                        if dist.javadocType == 'implementation':
-                            javadocArgs += ['--implementation']
-                        else:
-                            assert dist.javadocType == 'api'
-                        if dist.allowsJavadocWarnings:
-                            javadocArgs += ['--allow-warnings']
-                        javadoc(javadocArgs, includeDeps=False, mayBuild=False, quietForNoPackages=True)
                         tmpJavadocJar = tempfile.NamedTemporaryFile('w', suffix='.jar', delete=False)
                         tmpJavadocJar.close()
                         javadocPath = tmpJavadocJar.name
-                        emptyJavadoc = True
-                        with zipfile.ZipFile(javadocPath, 'w', compression=zipfile.ZIP_DEFLATED) as arc:
-                            javadocDir = join(tmpDir, 'javadoc')
-                            for (dirpath, _, filenames) in os.walk(javadocDir):
-                                for filename in filenames:
-                                    emptyJavadoc = False
-                                    src = join(dirpath, filename)
-                                    dst = os.path.relpath(src, javadocDir)
-                                    arc.write(src, dst)
-                        shutil.rmtree(tmpDir)
-                        if emptyJavadoc:
-                            if validateMetadata == 'full' and dist.suite.getMxCompatibility().validate_maven_javadoc():
-                                raise abort("Missing javadoc for {}".format(dist.name))
-                            javadocPath = None
-                            warn('Javadoc for {0} was empty'.format(dist.name))
+                        if getattr(dist, "noMavenJavadoc", False):
+                            with zipfile.ZipFile(javadocPath, 'w', compression=zipfile.ZIP_DEFLATED) as arc:
+                                arc.writestr("index.html", "<html><body>No Javadoc</body></html>")
+                        else:
+                            projects = [p for p in dist.archived_deps() if p.isJavaProject()]
+                            tmpDir = tempfile.mkdtemp(prefix='mx-javadoc')
+                            javadocArgs = ['--base', tmpDir, '--unified', '--projects', ','.join((p.name for p in projects))]
+                            if dist.javadocType == 'implementation':
+                                javadocArgs += ['--implementation']
+                            else:
+                                assert dist.javadocType == 'api'
+                            if dist.allowsJavadocWarnings:
+                                javadocArgs += ['--allow-warnings']
+                            javadoc(javadocArgs, includeDeps=False, mayBuild=False, quietForNoPackages=True)
+
+                            emptyJavadoc = True
+                            with zipfile.ZipFile(javadocPath, 'w', compression=zipfile.ZIP_DEFLATED) as arc:
+                                javadocDir = join(tmpDir, 'javadoc')
+                                for (dirpath, _, filenames) in os.walk(javadocDir):
+                                    for filename in filenames:
+                                        emptyJavadoc = False
+                                        src = join(dirpath, filename)
+                                        dst = os.path.relpath(src, javadocDir)
+                                        arc.write(src, dst)
+                            shutil.rmtree(tmpDir)
+                            if emptyJavadoc:
+                                os.unlink(javadocPath)
+                                if validateMetadata == 'full' and dist.suite.getMxCompatibility().validate_maven_javadoc():
+                                    raise abort("Missing javadoc for {}".format(dist.name))
+                                javadocPath = None
+                                warn('Javadoc for {0} was empty'.format(dist.name))
 
                     extraFiles = []
                     if deployMapFiles and dist.is_stripped():
